refactor(headlines_feed): remove initial fetch from UI
Removes the initial `HeadlinesFeedFetchRequested` event dispatch from the `initState` method of `HeadlinesFeedPage`.

This responsibility is being moved directly into the `HeadlinesFeedBloc`, which will now listen to the `AppBloc` to determine the correct moment to fetch data. This change eliminates a potential race condition and further decouples the UI from business logic orchestration.